requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.13
protobuf==3.6.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
